WIP DEMO: PATH not ignored with actions/setup-ruby
In this demo, we run the older actions/setup-ruby action,
then build a simple git-lfs.exe executable with Go, which
is in $PWD/build/git-lfs.exe.

If we set our $PATH environment variable to include the
$PWD/build path as the first path, and then run either
"git-lfs" or "git lfs", in both cases our demo executable
is found and run.  Note that in the second case, git.exe
is running first, and it locates our executable and runs it,
while in the first case we are running it directly.