WIP DEMO: PATH ignored with ruby/setup-ruby
In this demo, we run the newer ruby/setup-ruby action,
then build a simple git-lfs.exe executable with Go, which
is in $PWD/build/git-lfs.exe.

If we set our $PATH environment variable to include the
$PWD/build path as the first path, and then run "git-lfs",
our demo executable is found and run.

However, if we run "git lfs", the git.exe binary runs but
does not locate our git-lfs.exe; instead it runs the
default system git-lfs.exe, despite the fact that ours
should be located first according to the $PATH environment
variable.