t/testhelpers.sh: refine repo clone helpers
The clone_repo*() family of helper functions are used extensively
throughout our test suite; they generally run "git clone" and capture
its output, and then change the working directory to the newly cloned
repository and write the "git clone" output into a log file there
after first calling "git config".

The first introduction of a clone.log file in clone_repo() was in
commit bc025a8ced56f91ef791002b72b8dba2b0cb6337 of PR #679, and
similar log files are generated by the other functions that were
added later.

One consequence of the way these functions operate now is that,
if the "git clone" command should fail, then because we run all our
tests with the "set -e" option, no log file is generated.  The
function returns immediately, and as the log file is only created
at the end of the function, after changing directories, that step is
skipped.

A partial attempt to work around this behaviour was introduced in
commit 9a00eb3a684b61d33318bdfb89e3a1820b064c57 of PR #1893, but only
for the clone_repo_clientcert() function.  The "set +e" option is
temporarily set before running the "git clone" and then the exit code
captured before setting "set -e" again.  This allows the function to
grep for a specific error string in the log file which, if present,
indicates a problem that will not be treated as cause for immediate
failure.

Under most circumstances the current functionality in the other
functions is reasonable as any error messages "git clone" outputs
will appear in the stdout and stderr streams captured by our test
framework for debugging purposes.

However, this leaves no way for tests using these functions to
programmatically check the output of a failed "git clone" attempt.
For instance, if they are looking for a specific string to indicate
that the clone was unable to find a valid Git executable and ran
an invalid one instead (as per CVE-2022-24826), we may want to
be able to search clone.log even after a failure.

Therefore, in order to ensure that log traces are output and saved
in all cases, we change all of the clone_repo*() functions to pipe
the output of "git clone" into a "tee" command, which writes the
log file immediately.  Because the "tee" command is that last one
in the new command pipeline, this means any errors encountered by
"git clone" will not halt the function.

Instead, we capture the "git clone" exit code using PIPESTATUS[0],
and then return if it is non-zero.  If the clone was successful,
however, we proceed to change directory, move the log file into the
new repository (where many existing tests expect it to be available),
and then return after calling "git config".

In the clone_repo_clientcert() function specifically, this change
also allows us to remove the "set +e" temporary setting as it is
no longer useful, and reduce the number of places where we need
to write the log file.