docs/man/git-lfs-config.adoc: improve formatting
Our manual page for the "git lfs config" command has a number of minor
formattting inconsistencies and infelicities, such as not placing some
file paths and strings in backticks to ensure they are rendered by
Asciidoctor as literals, not using capitals for the term "URL",
so we refine those formatting details now.

Note that we now list all the accepted string values for Boolean
environment variable settings, inculding "true", "1", "on", "yes", and "t",
and the equivalent negative versions, in at least one or more of the
descriptions for these these types of variables.

We also update one Git documentation URL to contain the current
anchor tag for the relevant "http.<url>.*" section.