docs/man: improve formatting for various commands
As we expect to revise some of our commands' manual pages in subsequent
commits in this PR, we first address a number of minor formatting
inconsistencies and infelicities, such as not placing some file paths
and strings in backticks to ensure they are rendered by Asciidoctor
as literals, or using single quotes to distinguish a special term
where we would generally use double quotes.

Note that these changes are not an attempt at exhaustive copy-editing
or proofreading, merely a best effort to make the pages we expect to
revise further in this PR somewhat more consistent first.