t: add lfstest-genrandom command for tests
We use several techniques in our test suite to acquire random data
and filenames, of which our most prevelant one is to read from the
/dev/urandom special file.  Recently, this appears to no longer
work in our CI jobs on Windows GitHub Actions runners, possibly due
to an upstream change in Git Bash or for some other reason.

We also use the "uuidgen" command in one test in our t/t-prune.sh
test script to populate Git LFS object content.  However, that command
is not installed with Git Bash, so on Windows this data is simply empty;
while the test still passes, this is less than ideal.

We already have a number of commands written in Go to provide
platform-agnostic utilities to our test suite, and Go's crypto/rand
package can be used to generate random data on all the platforms
we support.  On Windows in particular it does not use /dev/urandom
(or a simulation of it, as currently depend on via Git Bash) but the
native Windows ProcessPrng API.

Therefore we add a new lfstest-genrandom command for our test suite
which uses the crypto/rand Go package to return strings of random data.
If a size is provided as a command-line argument, it returns that
number of random bytes; otherwise, it returns an effectively endless
stream of data.  This parallels the design of the Git test suite's
test-genrandom helper:

  https://github.com/git/git/blob/a7dae3bdc8b516d36f630b12bb01e853a667e0d9/t/helper/test-genrandom.c

When either the --base64 or --base64url command-line flags are specified,
the returned data is encoded either using the regular base64 specification
from RFC 4648, or using the alternate base64 specification (also from
RFC 4648) suitable for use in filenames and URLs.  In the latter case,
the slash and plus characters are avoided in favour of underscore and
dash characters.  In either case, the returned data has no base64 padding,
and if a size was provided on the command-line, will be truncated to
exactly that number of bytes after encoding (rather than being truncated
before encoding, which would produce extra output bytes).

While we could choose to only produce the URL-compatible form of base64
data, we would have to rewrite a number of tests which expect to be
able to read the random data we've written into a test file and then
pass it to our calc_oid() Bash function, which calls printf() on the
data.  If the data happens to begin with a dash character, printf()
would interpret it as a flag and could fail.

Conversely, if we only generated data encoded in the regular form of
base64, we would not be able to use it in the several places where
we generate random directory names without also passing it through
something like sha256sum, as we do now.  By making the URL-compatible
encoding available directly from our new utility command, we can
simplify these portions of our test suite when, in a subsequent commit
in this PR, we refactor our test's shell functions to make use of
our new command.