t/t-migrate-{ex,im}port.sh: avoid using mktemp
In commit 96775c62f71d795af1547f6ef92c2a8602fd32df of PR #2929 the
"migrate import (--object-map)" test was added to the test suite for
the "git lfs migrate import" command to validate the --object-map option
for that command, which was introduced in the same PR.  This test
makes use of the "mktemp" shell command to generate a unique temporary
directory, into which the test then writes several files including
the Git LFS migration's object map file.

Likewise, in commit 7632a205d378709bcf1d3733c31a0f43668c4821 of PR #3084,
a similar "migrate export (--object-map)" test was added to the test
suite for the "git lfs migrate export" command, which was introduced
in that PR, and this test also uses the "mktemp" shell command to
create a unique temporary directory.

However, neither test removes these temporary directories after
completion, so they remain in place indefinitely in whatever system
location is used by the "mktemp" command.  This is in addition to
the the temporary directory created by the "mktemp" command which
is used throughout all our test suites, whose path we store in the
GIT_LFS_TEST_DIR environment variable so it can be reused by all
test suites after the first one to run.

In a previous commit in this PR we introduced a new lfstest-genrandom
utility command for our test suite, and we can now make use of it to
replace both instances where our migration tests make an extra call
to the "mktemp" shell command (beyond the one used to populate the
GIT_LFS_TEST_DIR environment variable).

In both tests, we generate a unique file path under the path in the
GIT_LFS_TEST_DIR environment variable, using the --base64url flag with
our lfstest-genrandom command to ensure we have a safe directory name,
and then we create the directory.