t/testlib.sh: remove unused shutdown env var
The SHUTDOWN_LFS environment variable was added as part of the
original integration test suite in commit
b1f5025c7b3a1d98063aabf026fc21dddc41ba85 of PR #306, and was
used to indicate whether the Git server run as part of the test
suite should be stopped when the shutdown() shell function ran
at the end of a set of tests, or only when the last set of tests
completed.

When the integration test suite was refactored to use the "prove"
command and to align with Git's test suite, in PR #3125, most
references to the SHUTDOWN_LFS environment variable were removed,
particularly in commits 1926e942971a237c4c34e67e30fcd8a4b1d8244e
and  a4cc106eb3a4e9a282a5d21411327a01b9013dc9.

However, one reference was left in the t/testlib.sh script,
so we remove it now.