t/t-prune.sh: use lfstest-genrandom not uuidgen
In previous commits in this PR we introduced a new utility command
for our test suite, lfstest-genrandom, which reliably produces
random data on all the platforms we support (in several forms of
base64 encoding or as raw binary data), and then converted much of
our test suite to make use of this new command.

As noted in one of those commits, one test in our t/t-prune.sh test
script makes use of the "uuidgen" command to populate some Git LFS
objects with arbitrary text, but as this command does not exist in
the Git Bash environment on Windows, the blobs actually have empty
data in this case.  The test still passes because it doesn't depend
on the length of the random data, and the error exit code from the
missing "uuidgen" command is ignored.

We therefore revise this "prune verify large numbers of refs" test
(which was added in commit dd19c1b9c2d281b986829424ff57bd2bf3be2f29 of
PR #1861) to call our new lfstest-genrandom command instead of the
"uuidgen" command.  We use the --base64 flag and a slightly larger size
for the output base64-encoded string than the length of a UUID.

We also adjust the test so that it does not create the Git LFS objects
with sizes that do not match the length of their arbitrary content.
(Again, the test passed despite this inconsistency in its test data.)