t/{t-clone,testhelpers}.sh: drop macOS workaround
In commit 9a00eb3a684b61d33318bdfb89e3a1820b064c57 of PR #1893 we
introduced a workaround in some of our tests of Git LFS support for
TLS/SSL client certificates to deal with an exception raised when
using the version of libcurl installed by default by macOS, specifically
by macOS 10.12 ("Sierra"), which was the most recent release of macOS
at the time.

However, since at least v7.52.0, curl has included a fix for this
problem, and that will be included in all contemporary versions of
macOS.  By the time of macOS 10.14 ("Mojave"), the version of curl
included with the OS was v7.54.0, for instance.

We can therefore drop this workaround entirely, as it is no longer
applicable to any supported versions of macOS.

(Note, too, that our CI test suite always uses a libcurl version
installed using Homebrew.)