t/t-clone.sh: ignore TLS tests when using libnss3
Due to the issue described in #5658, the "cloneSSL" and "clone ClientCert"
tests in our t/t-clone.sh test script do not run to completion but exit
early, as a consequence of an improper check of the TRAVIS variable (which
is no longer used since we migrated our test suite to GitHub Actions in
PR #3808).

We expect to address this issue in subsequent commit in this PR, at which
point the tests will run fully, exposing a number of long-standing problems.
One of these occurs on the legacy CentOS 7 Linux system where the libcurl
library used by Git still depends on the libnss3 library, which has several
restrictions in the types of TLS certificates it accepts.  Specifically,
in our current builds on CentOS 7, libcurl version 7.29.0 is linked
against version 3.90 of libnss3.

The libnss3 library rejects certificates which have both Extended Key
Usage attributes for TLS Web Server Authentication, and the Basic
Constraint setting "CA:TRUE", meaning the certificate is intended
to be used as a Certificate Authority:

  https://security.stackexchange.com/questions/176177/why-does-curl-nss-encryption-library-not-allow-a-ca-with-the-extended-key-usage

However, certificates with both these settings are accepted by most
modern TLS libraries, and the self-signed certificate used by our
lfstest-gitserver utility program has both attributes.  Note that this
certificate is the one provided by the Go language's net/http/httptest
package, specifically the one from the net/http/internal/testcert package:

  https://github.com/golang/go/blob/7529d09a11496a77ccbffe245607fbd256200991/src/net/http/internal/testcert/testcert.go#L10-L33

Thus when we enable our "clone ClientCert" test, the second part of the
test, where it checks the use of an encrypted client key, will fail
on CentOS 7 because Git is dependent on the libnss3 library for TLS
certificate validation.

Moreover, we would also like to reverse the change made in commit
1cf7d6b131c3398fdc2197445427dae14b66cfcf of PR #1893, when the
"http.<url>.sslVerify" Git configuration option was set to "false"
in the setup() shell function of our t/testhelpers.sh library.  This
was done, according to the commit description, to try to get the
"clone ClientCert" test working in Travis CI.  We no longer use
Travis, as noted above, and we can fully enable this test with
"http.<url>.sslVerify" set to its default value of "true", at least
on the majority of the systems we use in our GitHub Actions CI jobs.

When we make this change as well, three tests in the t/t-clone.sh
script (the "cloneSSL", "clone ClientCert", and "clone ClientCert with
homedir certs" tests) will fail because libnss3 rejects the certifcate
provided by the Go language test server, for the reason described
above.  In all three tests, we see libnss3 return a "Certificate type not
approved for application" SEC_ERROR_INADEQUATE_CERT_TYPE error.

We expect to drop support for CentOS 7 soon, given that it has
reached the end of offical support in all but one distribution
(SUSE Linux Enterprise Server 12 SP5, for which general support
ends on October 31 of 2024).  As well, CentOS 8 and more recent
related distributions have switched away from building libcurl
against libnss3; in fact, support for that TLS library has been
dropped by the curl project entirely as of v8.3.0, per commit
curl/curl@7c8bae0d9c9b2dfeeb008b9a316117d7b9675175.

Therefore, rather than try to find a solution to this problem
by generating a different certificate for our lfstest-gitserver
utility program, we simply skip these three tests if we detect
that they are running on a Linux platform where the git-remote-https
binary depends on libnss3.  We can remove these exceptions later
once we drop support for CentOS 7.