t/t-clone.sh: use OpenSSL on Windows for TLS tests
Due to the issue described in #5658, the "cloneSSL" and "clone ClientCert"
tests in our t/t-clone.sh test script do not run to completion but exit
early, as a consequence of an improper check of the TRAVIS variable (which
is no longer used since we migrated our test suite to GitHub Actions in
PR #3808).

As well, since the "clone ClientCert with homedir certs" test was added
in commit c54c9da0735297c22cf52c1258e711fa76357686, as part of PR #5657,
it has skipped executing on Windows due to an error that was seen when
Git attempted to retrieve the passphrase for the encrypted key of the
TLS certificate used in that test from our git-credential-lfstest
helper program.

We expect to address the first issue in subsequent commit in this PR,
at which point the tests will run fully, exposing a number of
long-standing problems.

In the case of the "cloneSSL" test, Schannel's CertGetNameString() function
will reject the TLS certificate of our lfstest-gitserver program because
we connect to the server on the local IP address 127.0.0.1, and although
the certificate lists that address as an IPAddress field among its Subject
Alternative Name (SAN) attributes, the Schannel library only looks for
hostname matches with the DNSName fields and ignores any IPAddress fields.
As a result, Git returns a "failed to match connection hostname (127.0.0.1)
against server certificate names" error, as produced by libcurl.  As of
version 8.9.0 of libcurl, the CertGetNameString() function is called
with the CERT_NAME_DNS_TYPE option, which is documented to have this
behaviour:

  https://github.com/curl/curl/blob/5040f7e94cd01decbe7ba8fdacbf489182d503dc/lib/vtls/schannel_verify.c#L577-L582
  https://github.com/curl/curl/blob/5040f7e94cd01decbe7ba8fdacbf489182d503dc/lib/vtls/schannel_verify.c#L373-L378
  https://learn.microsoft.com/en-us/windows/win32/api/wincrypt/nf-wincrypt-certgetnamestringw#cert_name_dns_type

The TLS certificate presented by our lfstest-gitserver program is the
one provided by the Go language's net/http/httptest package, specifically
the one from the net/http/internal/testcert package:

  https://github.com/golang/go/blob/7529d09a11496a77ccbffe245607fbd256200991/src/net/http/internal/testcert/testcert.go#L10-L33

While it does include a DNSName field in its SAN list with the hostname
example.com, attempting to use this in our tests would involve forcing
both Git and Git LFS to override the available system DNS resolution.
For Git LFS in particular, there does not appear to be, at the moment,
a straightforward way of overriding the Go language's "net" package's
DNS resolution in a cross-platform and convenient manner.

Instead, with recent versions of Git and libcurl, we can bypass this
problem by setting the "http.sslBackend" Git configuration option
to "openssl", and thereby forcing the use of OpenSSL in place of
Schannel in libcurl.  This option has been available since commit
git/git@21084e84a418e0a6c4da72f632c5cd99386bd64b in Git version 2.20.0,
and is supported so long as the version of libcurl is at 7.56.0.
Since our CI jobs on Windows are executed on GitHub Actions where the
pre-installed Git for Windows releases are kept current, we can
make use of this option in our tests where TLS certificate valiation
is required.

Note that this option does not affect how Git LFS behaves, because
it uses the TLS validation provided by the Go net/http and crypto/tls
packages, and because the intent of our tests is to confirm correct
behaviour of Git LFS and not Git, using OpenSSL instead of Schannel
to ensure Git accepts our test server's TLS certificate on Windows
does not compromise the integrity of our tests.  (Moreover, having
the tests actually run to completion is a big step forward over the
current state where they pass in all circumstances due to the bug
in our conditional check of the TRAVIS variable.)

As for the problem observed when the "clone ClientCert with homedir certs"
test was added, Git reports the error "refusing to work with credential
missing host field" when libcurl uses the default of the Schannel
library on Windows, and so for this reason, the test has always
been skipped on Windows.  When we enable the "clone ClientCert" test
as well by removing the check on the TRAVIS variable, it too will
report the same error.

However, we can again resolve this problem by switching to the OpenSSL
backend, as noted a discussion on the Git mailing list:

  https://lore.kernel.org/git/TY0PR06MB54426B92E745B3035F0CC2DFD197A@TY0PR06MB5442.apcprd06.prod.outlook.com/

So for both of these tests as well, we set the "http.sslBackend"
Git configuration option to "openssl" when they run on Windows, and
we can then allow the "clone ClientCert with homedir certs" to
execute on that platform for the first time since it was introduced.