t/t-clone.sh: update progress meter message check
In commit db4e3b271d08da8110be0ee0ee091a2f9f212fbc of PR #2811 in 2018
we updated the progress meter text of Git LFS from "Git LFS" to the
more explicit "Uploading LFS objects" or "Downloading LFS objects".
A large number of were changed at this time to check for the revised
text strings in the output logs the capture from Git and Git LFS
operations.

However, the "cloneSSL" test in our t/t-clone.sh test script was not
updated because it does not actually run to completion, due to a
long-standing bug in the conditional used to check the TRAVIS
variable.  As described in #5658, the test instead always exits
early and declares success.

We expect to address this problem in a subsequent commit in this PR,
so we first update the test to check for the same progress meter
text as all our other tests, which will help ensure it succeeds
once it is fully enabled.

We also take the opportunity to add a similar check of the same
progress meter message to the "clone ClientCert" test.  This brings
our tests into slightly closer alignment with each other, which
should help ensure they stay consistent in the future.  (Note that
this test also does not run to completion at the moment, for the
same reason as the "cloneSSL" test.)