t/t-clone.sh: assert clean status when appropriate
In commit caaa0f91fad1afccc7c40197882160aea8b06d3c of PR #2647 and
number of tests of the "git lfs clone" and "git lfs pull" commands
were enhanced so as check that after those commands are invoked,
a "git status" command returns a "working tree clean" message.
To perform these checks, a call to our assert_clean_status() shell
function was added to the tests.

In the case of the "cloneSSL" test, an assert_clean_status() call
was added, but left commented out; it was then uncommented in commit
e0eede1602b09f1d3e81006ff7b30475f449b576 of the same PR.  Unfortunately,
the call is made when the current working directory has not yet been
changed to that of the newly cloned repository's working tree, and
so will fail as it stands now.

However, as described in #5658, the "cloneSSL" test and the "clone
ClientCert" tests in our t/t-clone.sh test script do not actually
run to completion, as a consequence of an improper check of the TRAVIS
variable (which is no longer used since we migrated our test suite to
GitHub Actions in PR #3808).  This bug was already in place at the time
of PR #2647, and so the assertions are never actually performed in
these tests.

We expect to address this problem in a subsequent commit in this PR,
and when we do so, the test "cloneSSL" test will fail because the
assert_clean_status() call is made without having changed the current
working directory to that of the cloned repository's working tree.

Therefore we move the assertion to the end of the block of checks
performed after the "pushd" shell command is used to change the
directory to that of the new clone's working tree.

We also take the opportunity to add the assert_clean_status() calls
to a number of other tests in the t/t-clone.sh script, so they are
all performing similar sets of checks.  This will help keep our
tests in closer aligment with each other.