t/t-clone.sh: check no improper "lfs" dir exists
In commit 365d6159898a438a7d26d305444614d349e8af9f of PR #1478 a
pair of tests in what is now our t/t-clone.sh test script were enhanced
to check that the "git lfs clone" command did not create a directory
named "lfs" within the top level ofa cloned repository's working tree,
as the "lfs" directory should instead be created within the ".git"
directory.

This check has since been replicated in a number of other tests
in the t/t-clone.sh script, but not all of them, so we add this
check in the relevant sections of the tests where it is not yet
performed.  We also adjust one check which dates from commit
365d6159898a438a7d26d305444614d349e8af9f to use the same -e shell
test as all the other checks, rather than the -f test.

These changes are not especially important, but in conjunction with
the revisions we have made in other commits in this PR, will help
synchonize the tests in the t/t-clone.sh script so they are more
consistent and complete.