t/t-clone.sh: indent all pushd/popd blocks
As we are making a number of changes to the tests in our t/t-clone.sh
test script in other commits in this PR, with the goal of increasing
the level of commonality between our tests, we take the opportunity
to indent all the code contained between "pushd" and "popd" commands.

This style of formatting is already used in a number of the tests,
but not all of them, so we make them consistent now, and we also remove
one line of excess whitepace.