t/testhelpers.sh: drop unused credential files
In commit c54c9da0735297c22cf52c1258e711fa76357686 of PR #5657 we add
support for the use of a client TLS certificate located in a user's home
directory, when specified using a tilde character in the path provided
for the "http.<url>.sslCert" and "http.<url>.sslKey" configuration
options.

As part of this change, we updated our setup_creds() shell test library
function to create two additional credential record files in the
directory named by our CREDSDIR environment variable.  These files
were intended to be read by our git-credential-lfstest helper program,
and contain the test passphrase for our client TLS certificate and key.

However, in practice, these files are never consulted by our test
credential helper, because the certificate and key are not encrypted and
we do not set either the "http.sslCertPasswordProtected" Git configuration
option or the GIT_SSL_CERT_PASSWORD_PROTECTED environment variable,
so neither Git nor OpenSSL search for a passphrase.

In particular, Git does not check with the configured credential helper
for a passphrase, so we don't need to provide credential record files
for our git-credential-lfstest helper program which correspond to the
filepaths of the client certificate and key files in the test user's
home directory.