t/{testhelpers,t-clone}.sh: local credential files
In commit b9602f3f3a95cd2b0ece154430025b6fd7eab89c of PR #5616 we
resolved a problem which caused flaky test results in our CI jobs
by ensuring that the t/t-credential.sh test script used a unique
directory for the credential record files we provide to our
git-credential-lfstest helper program.

This change ensured that when the "clone (HTTP server/proxy require
cookies)" test in our t/t-clone.sh script creates a record file named
"localhost" in the common directory specified by the CREDSDIR
environment variable, as a copy of the "127.0.0.1" file our
setup_creds() shell library function creates, this does not conflict
with the "credentials from netrc" test and other related tests
in the t/t-credential.sh test script, which depend on there not
existing a credential for the "localhost" hostname.

However, a previous attempt to resolve this problem was already
made in commits bb8794477042b14ed4bc1487189dbdefe71d3db0 and
bcb9eb3b0752727ce8152926fbeb5aa9f1355d94 of PR #3825, but assumed that
our test scripts would always run sequentially rather than in parallel.
In that PR, the "clone (HTTP server/proxy require cookies)" was
added, and at the end of the test, it removes the "localhost"
credential record file it creates at the beginning of the test.

Of course, if the test ever failed, then it would leave the "localhost"
file in place, so even if our test scripts always ran sequentially
there would still be a conflict with the tests in the t/t-credential.sh
script.

We can more comprehensively ensure that these two test scripts do
not conflict again in the future by using the same technique applied
in commit b9602f3f3a95cd2b0ece154430025b6fd7eab89c to the
t/t-credential.sh script in the t/t-clone.sh script as well.

We therefore add a call to setup_creds() at the start of the
t/t-clone.sh script, so it will also use a separate copy of the
credential record files in the CREDSDIR directory.  This means
that the "localhost" file will only exist in a credential record
directory used by the t/t-clone.sh script.  We can then also drop
the attempt to clean up the "localhost" file from the end of the
"clone (HTTP server/proxy require cookies)" test.

To make this change, we also need to slightly refactor our
t/testhelpers.sh library so that the setup_creds() function can
generate the appropriate credential record filenames on its own.
At the moment, it depends on some local variables defined in the
setup() function, but these are used nowhere else except in the
setup_creds() function, so we can just move their values directly
into the strings concatenated and passed to the write_creds_path()
function by the setup_creds() function.