t/t-clone.sh: use Git config for encrypted TLS key
We added support for encrypted TLS certificate keys in commit
706becaf20fc7657dcd296a17234c875f546603d of PR #3270, and as part of
that change, updated the "clone ClientCert" test in our t/t-clone.sh
test script to perform "git lfs clone" operations twice, once with
an unencrypted TLS certificate key and once with an encrypted one.

Because our test needs to ensure that Git does not defer to OpenSSL
to prompt the user for the passphrase to the encrypted key, but first
checks if the configured credential helper can provide the passphrase,
we set the GIT_SSL_CERT_PASSWORD_PROTECTED environment variable to "1".

However, we do this for both passes through the loop in our test,
so we set that environment variable when testing with the unencrypted
TLS key as well as with the encrypted one.  We only need to set it
in the latter case, though.

Moreover, setting the "http.sslCertPasswordProtected" Git configuration
value to "true" has the same effect as setting the environment variable
to "1".

To clarify the intent of the test's loops, we refactor the test now
to iterate over two pairs of values, with each pair containing the
TLS certificate private key file path to be used, and a "false" or
"true" value to be used for the "http.sslCertPasswordProtected" Git
configuration value.  This means that in the case of the unencrypted
key, we now avoid setting an environment variable or configuration
options which does not apply to that case.

We also take the opportunity to fix a small typo in the directory
name argument passed to the "git lfs clone" command, and we avoid
the need to delete the working tree created by the previous iteration
of the loop by including our loop index value in the directory name.