t/testhelpers.sh: allow TLS verification again
In PR #1893 we introduced support for TLS client certificates, and
while that was largely successful, some of our CI test jobs ran on the
Travis CI platform, and those encountered problems when validating the
TLS certificate generated by our lfstest-gitserver test utility program.
Several attempts were made to resolve this issue, and ultimately in
commit 1cf7d6b131c3398fdc2197445427dae14b66cfcf the "http.<url>.sslVerify"
Git configuration value was set to "false" to allow the "clone ClientCert"
test to pass in the Travis CI environment.

However, we no longer use the Travis CI platform, as we migrated our
CI jobs to GitHub Actions in PR #3808.  Further, in a prior commit in
this PR we have revised several tests in our t/t-clone.sh test script
so they do not run on the legacy CentOS 7 platform, where the libcurl
library used by Git depends on the libnss3 library.  The libnss3 library
does not accept the TLS certificate generated by our lfstest-gitserver
program, so these tests would otherwise fail on CentOS 7 if
"http.<url>.sslVerify" is set to "true".

As well, in another commit in this PR we have altered our t/t-clone.sh
test script to require the use of OpenSSL on Windows instead of the
default Schannel backend.

With these changes to our test suite, all our tests should now pass
if we allow the "http.<url>.sslVerify" Git configuration option to
default to "true", so we now remove the call which explicitly set
that option to "false".