t/{testlib,t-clone}.sh: quote Git config keys
As we have made a number of changes to our t/t-clone.sh test script in
previous commits in this PR, which reduced the number of "http.<url>.*"
Git configuration options for which we need to set distinct values,
we now revise the remaining such calls to "git config" to fully quote
the key names.

We also add quotation marks around the command substitutions we perform
to create the values we assign to several *GITSERVER shell variables in
our t/testlib.sh library.  These variables are then used throughout our
test suite as the leading portions of the Git remote URLs for our test
repositories.

These changes have no functional effect but help to keep our test code
more readable and consistent.

For a similar reason, we also adjust our "clone ClientCert with homedir
certs" test in our t/t-clone.sh script so that the initial setup sequence
of its test repository matches that of our other tests.