t: clarify non-Git env var exclusion tests
In commit 32244d92a44aaf24b144312bc1327ac7b4664d2e of PR #1825 the
APPVEYOR_REPO_COMMIT_MESSAGE variable was added to our CI test suite
environment in order to demonstrate that a problem exposed by the prior
commit 0951c698ad22a53c04d8d5d98a100f18f6e3993a in the same PR had been
addresssed.

Specifically, the earlier commit had the string "GIT_SSH" in its
message, and the AppVeyor CI service automatically provided the message
from the most recent commit in a Git branch under test in the
APPVEYOR_REPO_COMMIT_MESSAGE environment variable.  At the time,
the Environ() function in our "lfs" package would populate map it
returns with the names and values of any environment variables whose
names or values contained the string "GIT_".  The result was that
the APPVEYOR_REPO_COMMIT_MESSAGE variable was included in the returned
environment, which would cause the tests in what is now our t/t-env.sh
test script to fail.

The problem was resolved by adjusting the Environ() function to only
include environment variables in its map when their name contained
the string "GIT_".

To validate this fix, our test script environment was updated to
write a fixed value into the APPVEYOR_REPO_COMMIT_MESSAGE environment
variable (overriding whatever the AppVeyor service set) with an
example string containing the string "GIT_".  The example string used
was the commit message from the commit which exposed the problem.

Subsequently, we further revised the Environ() function in commit
f4f8fae03d5c945b328ecbe060c7d407d1982985 of PR #4269 so it only
includes environment variables whose names start with "GIT_",
excluding those whose names just include that string.

As we no longer use the AppVeyor service since we migrated our CI jobs
to GitHub Actions in PR #3808, and because the intent of the
APPVEYOR_REPO_COMMIT_MESSAGE variable is now somewhat obscure, we
replace it with a dedicated TEST_GIT_EXAMPLE environment variable
in our test scripts where the output of "git lfs env" is checked.
We ensure the "GIT_" string is included in both the name and value
of this variable, and we preface it with some comments to clarify
its purpose.

If a regression were ever to be introduced into the Environ() function
such that it did not exclude environment variables with "GIT_" in their
names or values, the tests in the t/t-env.sh and t/t-worktree.sh scripts
should still catch the problem.